edits based on comments

+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserService.java
@@ -9,13 +9,13 @@ import java.util.List;
 public interface UserService extends UserDetailsService {
     List<User> findAll();
 
-    User findById(int id);
+    User findById(long id);
 
     void save(User user);
 
-    void update(int id, User updatedUser);
+    void update(long id, User updatedUser);
 
-    void delete(int id);
+    void delete(long id);
 
     User findByUserName(String username);
